% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_c50.R
\name{print.stratified_rf}
\alias{print.stratified_rf}
\title{Print summary statistics from a model}
\usage{
\method{print}{stratified_rf}(x, ...)
}
\arguments{
\item{x}{A stratified_rf model.}

\item{...}{other options (not currently used)}
}
\description{
Print summary statistics from a model
}
\examples{
data(iris)
groups <- list(c("Sepal.Length","Sepal.Width"),c("Petal.Length","Petal.Width"))
mtry <- c(1,1)
m <- stratified_rf(iris,"Species",groups,mtry,ntrees=2,multicore=FALSE)
print(m)
}
