% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_c50.R
\name{summary.stratified_rf}
\alias{summary.stratified_rf}
\title{Summary statistics from a model}
\usage{
\method{summary}{stratified_rf}(object, ...)
}
\arguments{
\item{object}{A stratified_rf model.}

\item{...}{other options (not currently used)}
}
\description{
Summary statistics from a model
}
\examples{
data(iris)
groups <- list(c("Sepal.Length","Sepal.Width"),c("Petal.Length","Petal.Width"))
mtry <- c(1,1)
m <- stratified_rf(iris,"Species",groups,mtry,ntrees=2)
summary(m)
}
